Fixed coverage error
