--- .eslintrc.json
@@ -1,49 +1,0 @@
-{
-    "root": true,
-    "parser": "@typescript-eslint/parser",
-    "parserOptions": {
-        "ecmaVersion": 2020,
-        "sourceType": "module",
-        "ecmaFeatures": {
-            "jsx": true
-        }
-    },
-    "env": {
-        "browser": true,
-        "commonjs": true,
-        "es6": true
-    },
-    "settings": {
-        "react": {
-            "version": "detect"
-        }
-    },
-    "extends": [
-        "plugin:@typescript-eslint/recommended",
-        "plugin:prettier/recommended"
-    ],
-    "plugins": ["prettier", "@typescript-eslint"],
-    "rules": {
-        "@typescript-eslint/explicit-module-boundary-types": "warn",
-        "curly": ["warn", "multi"],
-        "quotes": [
-            2,
-            "double"
-        ]
-    },
-    "overrides": [
-        {
-            "files": [
-                "**/*.js"
-            ],
-            "excludedFiles": [
-                "node_modules",
-                "dist",
-                "build"
-            ],
-            "rules": {
-                "@typescript-eslint/no-var-requires": "off"
-            }
-        }
-    ]
-}
