+++ package.json
@@ -7,10 +7,18 @@
     "scripts": {
         "build": "tsc --build",
         "prepublishOnly": "pnpm run lint && pnpm run test && pnpm run build",
-        "test": "ts-mocha test/**/*.ts",
-        "lint": "eslint {src,test}/**/*.ts --fix",
-        "coverage": "nyc ts-mocha test/**/*.ts",
+        "test": "hayspec test",
+        "lint": "eslint {src,t2est}/**/*.ts --fix",
+        "coverage": "nyc pnpm run test",
         "prepare": "husky install"
+    },
+    "hayspec": {
+        "require": [
+            "ts-node/register"
+        ],
+        "match": [
+            "./test/**/*.test.*"
+        ]
     },
     "files": [
         "dist"
@@ -28,23 +36,23 @@
         "url": "https://github.com/throw-out-error/logger"
     },
     "devDependencies": {
-        "@types/chai": "4.2.16",
-        "@types/mocha": "8.2.2",
+        "@hayspec/cli": "^0.10.2",
+        "@hayspec/spec": "^0.10.2",
         "@types/node": "14.14.41",
         "@typescript-eslint/eslint-plugin": "4.22.0",
         "@typescript-eslint/parser": "4.22.0",
-        "chai": "4.3.4",
         "eslint": "7.24.0",
         "eslint-config-prettier": "8.2.0",
+        "eslint-import-resolver-typescript": "^2.4.0",
+        "eslint-plugin-import": "^2.22.1",
         "eslint-plugin-prettier": "3.4.0",
         "husky": "^6.0.0",
-        "mocha": "8.3.2",
         "nyc": "15.1.0",
         "prettier": "2.2.1",
-        "ts-mocha": "8.0.0",
+        "ts-node": "^9.1.1",
         "typescript": "4.2.4"
     },
     "dependencies": {
-        "ansi-styles": "^6.0.0"
+        "colorette": "^1.2.2"
     }
 }
